+++ packages/contract/lcov.info
@@ -1,113 +1,126 @@
 TN:
 SF:src/DaimoAccount.sol
-FN:86,DaimoAccount.initialize
+FN:87,DaimoAccount.initialize
 FNDA:4,DaimoAccount.initialize
-DA:91,4
 DA:92,4
-DA:94,4
-DA:95,0
-DA:98,4
+DA:93,4
+DA:95,4
+DA:96,0
 DA:99,4
-FN:103,DaimoAccount.executeBatch
+DA:100,4
+FN:104,DaimoAccount.executeBatch
 FNDA:0,DaimoAccount.executeBatch
-DA:104,0
 DA:105,0
-FN:134,DaimoAccount.validateUserOp
+DA:106,0
+FN:135,DaimoAccount.validateUserOp
 FNDA:0,DaimoAccount.validateUserOp
-DA:147,0
 DA:148,0
-FN:152,DaimoAccount.isValidSignature
+DA:149,0
+FN:153,DaimoAccount.isValidSignature
 FNDA:2,DaimoAccount.isValidSignature
-DA:156,2
-BRDA:156,0,0,1
-BRDA:156,0,1,1
-DA:157,1
-DA:159,1
-FN:166,DaimoAccount._validateSignature
+DA:160,2
+DA:162,2
+BRDA:162,0,0,1
+BRDA:162,0,1,1
+DA:163,1
+DA:165,1
+FN:172,DaimoAccount._validateSignature
 FNDA:2,DaimoAccount._validateSignature
-DA:173,2
-DA:176,2
 DA:177,2
-DA:180,2
+DA:178,2
 DA:181,2
-BRDA:181,1,0,-
-BRDA:181,1,1,-
 DA:182,2
+BRDA:182,1,0,-
+BRDA:182,1,1,-
 DA:183,2
-DA:186,2
+DA:184,2
 DA:187,2
 DA:188,2
-BRDA:188,2,0,-
-BRDA:188,2,1,-
-DA:190,2
-FN:194,DaimoAccount._validateUseropSignature
+DA:189,2
+BRDA:189,2,0,-
+BRDA:189,2,1,-
+DA:191,2
+FN:195,DaimoAccount._validateUseropSignature
 FNDA:0,DaimoAccount._validateUseropSignature
-DA:203,0
-BRDA:203,3,0,-
-BRDA:203,3,1,-
-DA:205,0
-DA:207,0
-FN:211,DaimoAccount._payPrefund
-FNDA:0,DaimoAccount._payPrefund
-DA:212,0
-BRDA:212,4,0,-
-BRDA:212,4,1,-
 DA:213,0
-FN:222,DaimoAccount._call
-FNDA:0,DaimoAccount._call
-DA:223,0
+DA:214,0
+DA:215,0
+DA:217,0
+DA:218,0
+BRDA:218,3,0,-
+BRDA:218,3,1,-
+DA:219,0
+DA:220,0
+DA:221,0
+DA:222,0
 DA:224,0
-BRDA:224,5,0,-
-BRDA:224,5,1,-
-FN:232,DaimoAccount._authorizeUpgrade
+DA:228,0
+DA:229,0
+BRDA:229,4,0,-
+BRDA:229,4,1,-
+DA:230,0
+DA:232,0
+FN:236,DaimoAccount._payPrefund
+FNDA:0,DaimoAccount._payPrefund
+DA:237,0
+BRDA:237,5,0,-
+BRDA:237,5,1,-
+DA:238,0
+FN:247,DaimoAccount._call
+FNDA:0,DaimoAccount._call
+DA:248,0
+DA:249,0
+BRDA:249,6,0,-
+BRDA:249,6,1,-
+FN:257,DaimoAccount._authorizeUpgrade
 FNDA:0,DaimoAccount._authorizeUpgrade
-FN:239,DaimoAccount.getActiveSigningKeys
+FN:264,DaimoAccount.getActiveSigningKeys
 FNDA:3,DaimoAccount.getActiveSigningKeys
-DA:247,3
-DA:248,3
-DA:249,3
-DA:250,3
-DA:251,768
-DA:252,768
-BRDA:252,6,0,768
-BRDA:252,6,1,4
-DA:253,4
-DA:254,4
-DA:255,4
-DA:258,3
-BRDA:258,7,0,-
-BRDA:258,7,1,-
-FN:264,DaimoAccount.addSigningKey
-FNDA:3,DaimoAccount.addSigningKey
-DA:265,3
-BRDA:265,8,0,-
-BRDA:265,8,1,-
-DA:266,3
-BRDA:266,9,0,-
-BRDA:266,9,1,-
-DA:267,2
-BRDA:267,10,0,-
-BRDA:267,10,1,-
-DA:268,3
-DA:269,3
-DA:270,2
-FN:275,DaimoAccount.removeSigningKey
-FNDA:4,DaimoAccount.removeSigningKey
-DA:276,4
-BRDA:276,11,0,-
-BRDA:276,11,1,-
-DA:277,3
-BRDA:277,12,0,-
-BRDA:277,12,1,-
-DA:278,2
+DA:272,3
+DA:273,3
+DA:274,3
+DA:275,3
+DA:276,768
+DA:277,768
+BRDA:277,7,0,768
+BRDA:277,7,1,4
+DA:278,4
 DA:279,4
 DA:280,4
-DA:281,2
+DA:283,3
+BRDA:283,8,0,-
+BRDA:283,8,1,-
+FN:289,DaimoAccount.addSigningKey
+FNDA:3,DaimoAccount.addSigningKey
+DA:290,3
+BRDA:290,9,0,-
+BRDA:290,9,1,-
+DA:291,3
+BRDA:291,10,0,-
+BRDA:291,10,1,-
+DA:292,2
+BRDA:292,11,0,-
+BRDA:292,11,1,-
+DA:293,3
+DA:294,3
+DA:295,2
+FN:300,DaimoAccount.removeSigningKey
+FNDA:4,DaimoAccount.removeSigningKey
+DA:301,4
+BRDA:301,12,0,-
+BRDA:301,12,1,-
+DA:302,3
+BRDA:302,13,0,-
+BRDA:302,13,1,-
+DA:303,2
+DA:304,4
+DA:305,4
+DA:306,2
 FNF:12
 FNH:6
-LF:53
+LF:64
 LH:41
-BRF:26
+BRF:28
 BRH:4
 end_of_record
 TN:
